refactor: type 분류별로 묶어서 하나의 파일에 지정
