+++ just-swim-app/src/app/_components/common/index.ts
@@ -1,4 +1,5 @@
 export * from './button';
 export * from './calendar';
 export * from './carousel';
+export * from './item';
 export * from './modal';
