+++ just-swim-app/src/app/_components/picker/datePicker/index.tsx
@@ -1,51 +1,16 @@
-'use client';
-
-import { HTMLAttributes } from 'react';
-
-import { Calendar } from '@components';
+import { Calendar, CalendarItem } from '@components';
+import { DatePickerProps } from '@types';
 
 import styled from './styles.module.scss';
 
 export function DatePicker({
   selectedDate,
   changeSelectedDate
-}: {
-  selectedDate: string,
-  changeSelectedDate: (date: string) => void,
-}) {
-  const DateBlock = ({
-    date,
-    isDisabled,
-    isToday,
-    isSelected,
-    clickCallback = () => {},
-    ...props
-  }: {
-    date: number,
-    isDisabled: boolean,
-    isToday: boolean,
-    isSelected: boolean,
-    clickCallback?: Function,
-  } & HTMLAttributes<HTMLButtonElement>) => {
-    const onClickButton = () => {
-      clickCallback();
-    }
-
-    return (
-      <button
-        className={`${styled.days} ${isDisabled ? styled.disabled : ''} ${isToday ? styled.today : ''} ${isSelected ? styled.selected : ''}`}
-        {...props}
-        onClick={onClickButton}
-      >
-        {date}
-      </button>
-    )
-  }
-
+}: DatePickerProps) {
   return (
     <div className={styled.date_picker}>
       <Calendar
-        DateBlock={DateBlock}
+        DateBlock={CalendarItem}
         selectedDate={selectedDate}
         changeSelectedDate={changeSelectedDate}
       />
