+++ just-swim-app/src/app/_components/picker/datePicker/styles.module.scss
@@ -1,31 +1,3 @@
-@import '@/globals.scss';
-
 .date_picker {
   padding: 20px;
-}
-
-.days {
-  width: 100%;
-  height: 34px;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  color: black;
-}
-
-.days.disabled {
-  color: $--gray-300;
-  cursor: default;
-}
-
-.days.today {
-  color: $--Primary-blue;
-}
-
-.days.selected {
-  background-color: $--Primary-blue;
-  border-radius: 50%;
-  color: white;
 }
