--- just-swim-app/src/app/_types/button/ConfirmButtonProps.ts
@@ -1,7 +1,0 @@
-export interface ConfirmButtonProps {
-  text: string;
-  kind: 'confirm' | 'confirm-sub' | 'cancel' | 'cancel-sub' | 'normal';
-  border?: boolean;
-  loading?: string;
-  active?: boolean;
-}
