--- just-swim-app/src/app/_types/button/index.ts
@@ -1,1 +1,0 @@
-export * from './ConfirmButtonProps';
