--- just-swim-app/src/app/_types/form/ColorInputProps.ts
@@ -1,4 +1,0 @@
-export interface ColorInputProps {
-  name: string;
-  defaultValue?: string;
-}
