--- just-swim-app/src/app/_types/form/DateInputProps.ts
@@ -1,7 +1,0 @@
-export interface DateInputProps {
-  name: string;
-  valid?: boolean;
-  defaultValue?: string;
-  renderIcon?: Function;
-  suffix?: string;
-}
