--- just-swim-app/src/app/_types/form/DayInputProps.ts
@@ -1,5 +1,0 @@
-export interface DayInputProps {
-  name: string;
-  valid?: boolean;
-  defaultValue?: string;
-}
