--- just-swim-app/src/app/_types/form/FileInputProps.ts
@@ -1,5 +1,0 @@
-export interface FileInputProps {
-  name: string;
-  length?: number;
-  size?: number;
-}
