--- just-swim-app/src/app/_types/form/FormButtonProps.ts
@@ -1,5 +1,0 @@
-export interface FormButtonProps {
-  text: string;
-  loading?: string;
-  active?: boolean;
-}
