--- just-swim-app/src/app/_types/form/LocationInputPros.ts
@@ -1,4 +1,0 @@
-export interface LocationInputPros {
-  name: string;
-  valid?: boolean;
-}
