--- just-swim-app/src/app/_types/form/TextAreaProps.ts
@@ -1,7 +1,0 @@
-import { ChangeEvent } from "react";
-
-export interface TextAreaProps {
-  name: string;
-  height?: number;
-  onChange?: (event: ChangeEvent<HTMLTextAreaElement>) => void;
-}
