--- just-swim-app/src/app/_types/form/TextInputProps.ts
@@ -1,4 +1,0 @@
-export interface TextInputProps {
-  name: string;
-  valid?: boolean;
-}
