--- just-swim-app/src/app/_types/form/TimeInputProps.ts
@@ -1,6 +1,0 @@
-export interface TimeInputProps {
-  name: string;
-  valid?: boolean;
-  defaultValue?: string;
-  defaultTimeValue?: string;
-}
