--- just-swim-app/src/app/_types/form/index.ts
@@ -1,9 +1,0 @@
-export * from './ColorInputProps';
-export * from './DateInputProps';
-export * from './DayInputProps';
-export * from './FileInputProps';
-export * from './FormButtonProps';
-export * from './LocationInputPros';
-export * from './TextAreaProps';
-export * from './TextInputProps';
-export * from './TimeInputProps';
