+++ just-swim-app/src/app/_types/typeCommon.ts
@@ -1,0 +1,35 @@
+import { Dispatch, MouseEvent, SetStateAction } from "react";
+
+export interface ConfirmButtonProps {
+  text: string;
+  kind: 'confirm' | 'confirm-sub' | 'cancel' | 'cancel-sub' | 'normal';
+  border?: boolean;
+  loading?: string;
+  active?: boolean;
+}
+
+export interface CalendarProps {
+  DateBlock?: JSX.ElementType,
+  selectedDate?: string,
+  changeSelectedDate?: (date: string) => void
+}
+
+export interface ImageCarouselProps {
+  images: string[],
+  index: number,
+  setIndex: Dispatch<SetStateAction<number>>,
+  useDeleteButton?: boolean,
+  deleteImage?: (index: number) => void,
+  hideModal: (event: MouseEvent<HTMLButtonElement>) => void,
+}
+
+export interface ModalBodyProps {
+  children?: React.ReactNode,
+  hideModal: (event: MouseEvent<HTMLButtonElement>) => void,
+}
+
+export interface ConfirmModalProps {
+  children?: React.ReactNode,
+  hideModal: (event: MouseEvent<HTMLButtonElement>) => void,
+  confirmCallback: (event: MouseEvent<HTMLButtonElement>) => void,
+}
