+++ just-swim-app/src/app/_types/typeConstants.ts
@@ -1,0 +1,11 @@
+export interface DayProps {
+  "monday": boolean,
+  "tuesday": boolean,
+  "wednesday": boolean,
+  "thursday": boolean,
+  "friday": boolean,
+  "saturday": boolean,
+  "sunday": boolean,
+  [props: string]: any,
+}
+
