+++ just-swim-app/src/app/_types/typeLayout.ts
@@ -1,0 +1,5 @@
+export interface HistoryBackHeaerProps {
+  title: string,
+  additionalLink?: string,
+  additionalContent?: string,
+}
