+++ just-swim-app/src/app/_types/typeModal.ts
@@ -1,0 +1,27 @@
+import { MouseEvent } from "react";
+
+import { DayProps } from "@types";
+
+export interface ColorModalProps {
+  initialColor: string,
+  hideModal: (event: MouseEvent<HTMLButtonElement>) => void,
+  setColor: (color: string) => void,
+}
+
+export interface DateModalProps {
+  initialDate: string,
+  hideModal: (event: MouseEvent<HTMLButtonElement>) => void,
+  setDate: (color: string) => void,
+}
+
+export interface DayModalProps {
+  initialDays: DayProps,
+  hideModal: (event: MouseEvent<HTMLButtonElement>) => void,
+  setDays: (days: DayProps) => void,
+}
+
+export interface TimeModalProps {
+  timeValue: string,
+  hideModal: (event: MouseEvent<HTMLButtonElement>) => void,
+  setTimeValue: (time: string) => void,
+}
