+++ just-swim-app/src/app/_types/typePicker.ts
@@ -1,0 +1,17 @@
+export interface ColorPickerProps {
+  selected: string,
+  change: (color: string) => void,
+}
+
+export interface DatePickerProps {
+  selectedDate: string,
+  changeSelectedDate: (date: string) => void,
+}
+
+export interface TimePickerProps {
+  value: string,
+  setValue: (time: string) => void,
+  itemHeight?: number,
+  itemsToShow?: number,
+  paddingY?: number,
+}
